Split out Huton from the NIN Combos module
